fix(consensus): Correct several artifact bounds and update docs (#1074)
We had an off-by-one error in the upper bound for finalizations. This
revealed a few other inaccuracies in our calculations, mostly due to
confusion about inclusive ranges.

This PR updates the illustration of our worst-case scenario, adjusts the
parameter `l` to reflect total number of blocks in the relevant range,
and includes some details that were previously not accounted for
(deletion of shares *at* finalized height, block maker producing
additional block).