test: Subnet memory state machine test (#935)
This introduces a new canister called `random-traffic-test-canister`
that generates random payloads and sends it to random receivers (out of
a provided list). It uses random numbers generated internally, but the
seed can be provided from outside to ensure determinism. Additionally it
operates on a weighted binomial distribution to make random downstream
calls; this can be disabled by setting the weight for downstream calls
to 0.

Depending on the list of receivers and the number of canisters it can be
used to simulate random bi-directional traffic or a client - server
model where only one side makes calls and so forth.

It keeps track of the calls it makes in internal 'records' which can be
queried in order to precisely understand what kind of traffic was
actually simulated, how big the payloads were, whether calls were
rejected and so forth.

The canister is intended for use in `StateMachine` tests. Along with the
canister is a framework to write such tests and a proof of concept test
that checks the memory used for guaranteed response messages does not
exceed the configured limit. I expect to use this canister for more
tests such as under canister migration and for testing reject signals
for requests following this PR.

---------

Co-authored-by: Alin Sinpalean <58422065+alin-at-dfinity@users.noreply.github.com>
Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>