refactor(nns): Migrate nns governance from dfn_core -> ic_cdk (#1565)
We are migrating from using dfn_core (which is unmaintained) to ic_cdk
(which is maintained, and has features unavailable in dfn_core).

The major difficulty was previously the difference in behavior with
dfn_core::call and ic_cdk::call as it relates to calling drop on
variables in memory. We relied on the lack of cleaning up in
dfn_core::call for LedgerUpdateLock to make sure that a subsequent panic
(after a call to the ledger) would keep the neuron in a locked state,
and not leave the neuron unlocked but in an inconsistent state. To fix
this, we rely on a new function available in ic_cdk,
'is_recovering_from_trap' which tells us if there was a panic.

Additionally, we migrate all of the macros to use ic_cdk, which removes
most of the need for our previous export and rename strategy with candid
(i.e. using 'over' and 'over_async' and then calling into the method
that has the signature we export into the candid file)This is possible
because ic_cdk automatically wraps all the methods appropriately.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>