feat(EXC-1716): Canister Backtrace on Trap (#1449)
Resolves EXC-1716. We introduce an optional field in the
`Hypervisor::Trapped` error variant to contain a backtrace captured from
the canister when it traps. The field will be `None` if the
`canister_backtrace` feature is not implemented. Otherwise it will be
`None` if the canister doesn't include a "name" custom section or the
calling canister doesn't have permissions to view the backtrace.

Follow up changes will capture the backtrace in other error situations,
dump it to the canister log, and strip it out of reject responses when
the caller doesn't have sufficient permissions.