feat(call-v3): [NET-1753] Return a certificate for duplicate requests that are already executed (#1523)
The replica node will currently time out if a duplicate messageis
submited, if the message was already executed. This PR adds a check to
the call handler, to check if the message is already in the certified
state. If the message is indeed in the certified state, then there is no
need for the call handler to submit the message to the ingress pool.

closes NET-1753