chore(PocketIC): route requests to /_/ in PocketIC HTTP gateway (#1574)
The PocketIC HTTP gateway routes requests whose paths start with `/_/`
and for which no canister ID can be found directly to the PocketIC
instance/replica (this only used to apply to requests for `/_/dashboard`
independently of whether a canister ID could be found). This way, the
PocketIC HTTP gateway could be used to browse all "hidden" replica paths
in local development without talking to the replica directly (which
requires distributing two URLs to the client - one for the gateway and
another one for the replica).