fix: Defragment correct state (#1683)
This is a follow-up to commit a438bb7. The idea of the previous commit
was to defragment the memory layout of the list of canisters. This fix
makes sure we actually defragment the correct state, namely the one that
has been fragmented in earlier rounds. The previous PR was perfectly
safe, but did not provide the desired performance benefit.