refactor(sns): Remove different sns init behavior when cfg(feature = "test") (#1609)
[Next PR →](https://github.com/dfinity/ic/pull/1606)

## Problem

ic-sns-init has :init and :init--test-feature targets, which makes our
dependency situation very hairy due to the viral nature of test
features. It should be removed, but first ic-sns-init needs to not have
any differential behavior between the test and normal configurations.

## Solution

Remove the code that ran when cfg(feature = "test") was true. This was
code that set up ledger with a principal with a ton of tokens.

This was probably needed at first, but is no longer, because the test
configuration of the SNS Gov canister has a `mint_tokens` API that can
be used instead. If this breaks anyone, I will help them switch over to
the new API. I think the `nns_dapp` has already switched to the new API.