refactor(IDX): switch from title to labels (#1675)
This PR switches from using PR title logic for specific behavior to
using labels. Some key benefits are:
- You won't have to change PR titles before merging (or forget to)
- You can select from pre-existing labels and there is less risk for
typos and confusion

Additionally, this simplifies the bazel `main.sh` and `run-build-ic.sh`
logic, as we only need to check if `RUN_ON_DIFF_ONLY` was set and not
check the PR title.

The next PR will update the logic for `override-didc-check`.

---------

Co-authored-by: IDX GitHub Automation <IDX GitHub Automation>
Co-authored-by: Nicolas Mattia <nicolas.mattia@dfinity.org>