feat: [NODE-1469] Change monitoring strategy for GuestOS VM (#1586)
In 24.04 systemd is getting more strict with the use of `PIDFile`, and
no longer allows arbitrary monitoring by enforcing correct cgroups.

We've been managing the GuestOS VM with this mechanism. Instead, this
will have virsh manage the liveness of the VM with:
https://github.com/dfinity/ic/blob/f7e561c00a2745f946372f5166fd7968fa664f53/ic-os/components/hostos-scripts/guestos/guestos.xml.template#L31-L33

and our service will only control starting and stopping so that we can
still schedule around GuestOS launch.