fix(NODE-1483): query_nns_nodes bug (#1665)
[NODE-1483](https://dfinity.atlassian.net/browse/NODE-1483)

Two bugs exist in assemble_nns_nodes_list and fetch_property. In
assemble_nns_nodes_list, fetch-property is called without passing a
metric value, which causes the call to fail to populate NNS_URL_STRING.

Then, when query_nns_nodes was called, the for loop was never entered
and no failure was caught.


This PR updates fetch-property to make the metric field optional if
metrics.sh does not exist and it also refactors query_nns_nodes to fail
if NNS_URL_LIST is empty.

I tested this change on bare metal 👍 

[NODE-1483]:
https://dfinity.atlassian.net/browse/NODE-1483?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ