feat: Introduce p2p slot table limit and limit allowed ingress slots per peer (#1213)
To support hashes in blocks there are modifications #1061 to the ingress
pool that would allow the unvalidated ingress pool to grow without
bounds. P2P allows for bounding the unvalidated pool via the slot table.
Until now the slot table didn't have any bounds and this MR introduces a
config option to bound the slot table and introduces a hard bound for
the ingress pool of `50000`. Other clients use `usize::MAX` as a bound.