fix: use `OnceCell` to store nns certificate delegation and use it in https outcalls transform function (#875)
Currently the certificate delegation is only used in the http server
because it is needed for the query/read-state request. The certifiacte
is obtained by calling the status endpoint of a NNS node on startup.

This fixes a bug in the https outcalls transform function that no
certificate is passed to the callback. To be able to use the certificate
in other places we use a `OnceCell` that is shared between consumers of
the delegation and which is populated by the http endpoint crate on
startup.