feat(IDX): use json for mainnet-canisters
This replaces `mainnet-canisters.bzl` with a json equivalent
`mainnet-canisters.json`. This simplifies the update logic and separates
the bazel logic from the mainnet data, also allowing mainnet data to be
used more broadly if necessary. In the future this will allow us to
replace the update logic with `jq` and `ic-wasm` directly.

A new repository rule is added which reads the JSON data and downloads
the canisters as repositories.