feat(nervous-system): Filter out boring commits when listing new NNS and SNS commits. (#1647)
Overall, tries to make list-new-commits.sh more ergonomic.


# Ergonomics

Boring is defined as having one of the following Conventional Commits
prefixes:

- chore
- refactor
- test

Colorizes.

When commit ID is not supplied, warns that one is chosen automatically.
(When I originally sent this out for review, commit ID was required, but
feedback said this is too inconvenient. So, I switched to printing a
warning.)


# Entrained Change(s)

Factored out common code from two loops (one for NNS, the other for
SNS).

Adds more print_* colors.


# Possible Future Work (no promises)

Instead of manually reviewing which canisters to upgrade, this could be
used to make that choice for us (using the same criterion for using
green to print the canister heading). This would help streamline our
releases, and in particular, reduce the release tsar's workload.