refactor(NODE-1484): refactor os_tools and networking code (#1666)
NODE-1484

Move the mgmt_mac conditional from generate_mac_address to the os_tool
top level. This reduces the responsibility of generate_network_config
and this change means we aren't passing optionals far down the stack
(mgmt_mac, deployment_name). And this change makes it more clear that
mgmt_mac is a branch for testing logic.