feat(nns): More secure random numbers. (#1633)
In NNS Governance.

This is done by reseeding every hour. The seed is sourced from the
raw_rand method of the management canister, and because of this, these
seeds are "as secure as it gets". This is a compromise between speed and
security that DFINITY's security team is ok with.

Currently, there is no need for secure RNs. However, that could come up
in the future, and when it does, it would be too easy for us to overlook
(without this change).

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>