chore: [MR-616] Drop `CanisterQueue::QueueItem` proto, part 1
Instead of using enum references, as initially planned, we use plain numeric references but look them up in the pool plus a couple of compact reject response maps. So the `CanisterQueue::QueueItem` proto is no longer necessary.

As a first step, add and also populate a new, simpler representation of canister queues. In a second step (and after a replica release) we can then drop the old representation and the `CanisterQueue::QueueItem` message.