chore: [MR-616] Drop `CanisterQueue::QueueItem` proto, part 1 (#1797)
[MR-616]: Instead of using enum references, as initially planned, we use
plain numeric references but look them up in the pool plus a couple of
compact reject response maps. So the `CanisterQueue::QueueItem` proto is
no longer necessary.

As a first step, add and populate a new, simpler representation of
canister queues (using plain integers). In a second step (and after a
replica release) we can then drop the old representation and the
`CanisterQueue::QueueItem` message.

[MR-616]:
https://dfinity.atlassian.net/browse/MR-616?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ