Allow for setup to finish inside SetupOS when booted inside QEMU.
The setup routine uses /dev/vda to set up HostOS and GuestOS.
Unfortunately, that drive contains the setup partition, which
is happily destroyed by the installer as the installer proceeds.
When time comes to reboot, a bunch of gobbledygook is printed
to the console, and the system just hangs uselessly after that.

To work around that issue and permit installation of the OS in the
QEMU environment, we create an empty disk (sparse) and use that as
the first disk during QEMU invocation.  This solves the issue of
SetupOS destroying itself prior to reboot.

We also increase startup performance by avoiding the copy of the
tarball that contains the SetupOS image, and by asking tar to be
sparse with data copies whenever it can be.

Finally, we attempt to clean up the folder upon completion of the
execution.  This avoids large files remaining around after setup.

It's truly space-efficient eh!

```
manuel@devenv-manuel:~$ du -h /tmp/tmp.FO1kRm0gqF.qemu-launch-remote-vm/*
2.5G    /tmp/tmp.FO1kRm0gqF.qemu-launch-remote-vm/disk.img
1.6G    /tmp/tmp.FO1kRm0gqF.qemu-launch-remote-vm/target.img
```