feat: RUN-1038: Propagate execution mode (wasm64/32) to replica (#1784)
This PR adds functionality to make the replica aware of the
Wasm32/Wasm64 execution mode of canisters. This will be used during the
execution phase to track cycle consumption separately.

To have this, the canister execution state adds a `is_wasm64` flag. This
flag is computed by the compiler sandbox the first time a canister is
installed, or when it is upgraded.