refactor(sns): Use ic-nervous-system-agent in sns-audit (#1756)
## Problem

sns-audit is available as a library, which makes it tempting to use in
tests and in the CLI. However, there are some blockers:

1. the sns-audit library requires ic-agent, which is not usable in the
context of pocketic tests as far as I know.
2. I would like new CLI functionality to be written using
ic-nervous-system-agent as it makes the code much more readable.

Additionally, the errors from sns-audit errors are all strings. This is
not optimal for consuming it as a library where you might want to have
some logic that depends on the particular error that occurred, if one
didi.

## Solution

Rewrite sns-audit to use ic-nervous-system-agent, and use an explicit
error type rather than using strings.

## Notes

The error messages have changed slightly. They no longer mention the
SNS's name. It is now mentioned sooner.

New:
```
❯ bazel run :sns-audit --config=lint "https://ic0.app/" "mlqf6-nyaaa-aaaaq-aadxq-cai"                                                          
sns_name = Juno Build
sns_tokens_per_icp = 549.4149218088792154581282989
Error: swap is not in the final state yet, so `initial_neurons_fund_participation` is not specified.
```

Old:
```
❯ bazel run :sns-audit --config=lint "https://ic0.app/" "mlqf6-nyaaa-aaaaq-aadxq-cai"
sns_tokens_per_icp = 549.4149218088792154581282989
Error: "SNS swap Juno Build is not in the final state yet, so `neurons_fund_refunds` is not specified."
```

This was done to make the error type simpler. Otherwise, it would have
to carry the SNS name information in all cases where it was relevant.
Since it is still printed slightly higher up, and hopefully people know
the SNSes they are auditing anyway, I hope that's just as good.

## How to review this PR

The first commit makes changes to ic-nervous-system-agent to add helpers
that will be useful to sns-audit.

The second commit changes sns-audit to use ic-nervous-system-agent.
There, you can see that the implementation of sns-audit is dramatically
simplified – it now is not concerned with the details of how to call the
canisters, as that is handled by ic-nervous-system-agent.