fix: [MR-523] Prevent duplicates of aborted or paused responses (#1851)
[MR-523]: While a response is in DTS execution, we no longer have a
response with that `CallbackId` in `CanisterQueues` (meaning we wouldn't
detect it as a duplicate callback); but we still have the callback
itself (so that a second response would match an existing callback).

Explicitly check against the `CallbackId` of a potential aborted or
paused response in `SystemState::push_input()` and
`SystemState::induct_messages_to_self()`.

[MR-523]:
https://dfinity.atlassian.net/browse/MR-523?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ