refactor: [EXC-1724] Refactor task_queue (#1490)
Refactor SystemState::task_queue to be a new struct with an additional
field on_low_wasm_memory_hook_status, to ensure that whenever available
first task that will be poped from the task queue is OnLowWasmMemoryHook
(excluding paused executions).

In the [follow-up](https://dfinity.atlassian.net/browse/EXC-1725) we
will add logic in which when doing operations on task queue, in the case
there are no paused executions in the queue, it is checked if
`on_low_wasm_memory_hook_status` is `Ready`, and if so it will be
returned as it is on the front of the queue. That will ensure that the
hook is executed right after the execution of the message/task that
triggered the hook condition.

---------

Co-authored-by: Alin Sinpalean <58422065+alin-at-dfinity@users.noreply.github.com>
Co-authored-by: Dimitris Sarlis <dimitrios.sarlis@dfinity.org>