refactor: Encapsulate the CallContextManager within SystemState (#1498)
Make `SystemState::status` private and get rid of all methods that
expose mutable references to either the `CallContextManager` or to
individual call contexts.

While this by itself does not ensure perfect consistency between
`CanisterQueues` and `CallContextManager` (messages are still pushed and
popped independently of registering / unregistering call contexts and
callbacks; and DTS execution actively requires that a callback hang
around until after the response was executed), it does allow for greater
control; and for adding some internal consistency checks in the future.

Maybe even pave the way for formalizing e.g. that the execution of a
specific response / callback / call context is in progress (so the
respective callback may still be in place without a corresponding
response in the queues; but everything else should match up).