test(IDX): reduce nightly system-test load (#1898)
Farm is overloaded whenever the bazel-system-test-nightly workflow is
running. It looks like this is causing this workflow to [fail
often](https://github.com/dfinity/ic/actions/workflows/schedule-daily.yml?query=branch%3Amaster).

We have many system-tests that are configured to run both on hourly and
nightly. To reduce load we remove the system-tests from nightly that are
also configured to run hourly.

Additionally there were a number of nightly-only system-tests that use
the `system_test_nns` macro which caused two variants of the test to run
on nightly: one using the mainnet NNS and one using the HEAD NNS. This
PR turns the latter tests into manual. We recognise this reduces test
coverage so component teams have to review carefully if this trade-off
is acceptable.