feat(IC-1579): TLA-code link for split_neuron (#913)
This is the first change to link the behavior in the existing tests of
the `split_neuron` method on the NNS governance with the corresponding
TLA model. In this PR, only the pure Rust-based tests are instrumented
to check the link; I plan to follow up with another PR linking the
canister-based tests as well (this will require adding a new test-only
query method to fetch the TLA traces), but I'm opening this now to allow
Andy to work on `claim_neuron` concurrently.

The PR consists of roughly four sets of changes:
1. Conditional (under the `tla` crate feature) instrumentation of
production code
2. TLA model of split neuron and the mapping of the Rust state to TLA
state (the `tla` module under governance)
3. Conditional checks in the Rust based tests of split neuron. These can
be disabled permanently through `feature_flags.bzl` or on the
temporarily on the Bazel command line with `--define tla_disabled=true`
4. The Bazel/Cargo config. There's perhaps a bit more splitting of Bazel
targets into "regular" and `--test_feature` than optimal, due to the
fact that the SNS `swap` depends on the `IcpLedgerCanister` even when
not compiled into a canister, but I didn't dare try to change this.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>