chore: Revert risky changes for load issues (#1936)
This PR reverts all the changes we consider risky for master. The idea
is to have a clean RC cut with safe code from master (although without
fixing mainnet issues). We will continue to work on some tweak on the
fixes on the side and we'll create another RC (potentially hotfix) that
includes these changes again and any tweaks when we're confident that
they will perform well on mainnet.