chore: Stop using NeuronId for proposal IDs (#1950)
# Motivation

NNS Candid files used to be generated from proto files. Because of this
the structurally identical `ProposalId` and `NeuronId` types go
conflated into just `NeuronId`.

Now we no longer generate the Candid files from the proto files so we
can fix the types.

# Change

1. Add `ProposalId` type in `governance.did` and `governance_test.did`.
2. Use `ProposalId` type where appropriate.

# Tested

Not tested but the existing interface test should pass.