chore: Change how the build time is calculated (#1876)
Previously, we computed the build time when building the SetupOS base
image. We're trying to move away from base images and want to be able to
build the entire OS directly from source, therefore a new solution is
required for reproducibility which does not take the current time into
account. The new approach uses the commit timestamp.

In order to allow better caching of non-release builds, a new build flag
`release_build` is introduced. Non-release builds use a constant commit
time.

This approach assumes/requires that commits are fairly regular. We
currently check that the build time (or commit time after this change)
is not older than 6 weeks.