feat(nns): Heap neuron count metric. (#1931)
# Rationale

Not that justification is needed for more stats, but there was already
an analogous metric for stable memory.

# Entrained Change(s)

Added headings to the `encode_metrics` function, because it's very long.

Moved a couple of metrics around so that they are under the right
heading.

# Note to Reviewers

It would probably be easier to step through the commits in this PR.
Otherwise, if you view all the changes at once, they will be all mixed
together, and probably more confusing than just stepping through the
commits.

---------

Co-authored-by: Arshavir Ter-Gabrielyan <arshavir.ter.gabrielyan@dfinity.org>