test(IDX): enable IPv4 and A records for PrometheusVms (#1977)
Some folks have problems accessing the Grafana and P8s server in
testnets via the registered domain name because by default only a AAAA
record is registered. In some networks (in particular the DFINITY
office) AAAA records don't work well. We therefor enable IPv4 support
for PrometheusVms by default and when enabled register an A record for
the prometheus and grafana server.