feat(sns): Add Swap.reset_timers (#1989)
Now that the Swap canister's periodic tasks are migrated from heartbeats
to timers, a theoretical risk can arise if a panic occurs during one of
the timers' setup process. This could occur, e.g., if the Swap runs out
of cycles during that code's execution. To be able to recover from such
a situation, the NNS could already pass two proposals (one to publish a
new Swap WASM, potentially identical to the one already installed,
thereby creating a new step in the upgrade order; another one for
triggering the Swap upgrade). However, it would be better if the SNS
could recover itself. To that end, we add a new function for resetting
Swap's timers, only callable by the SNS Governance. A follow-up PR will
add an SNS proposal for calling this new function.

< [Previous PR](https://github.com/dfinity/ic/pull/1932) |