fix: integer overflow in NNS tests in `is_inactive`  (#1984)
In tests, `now` was observed to be set to something between 50 and 60
seconds. When `now` is set to less than 2 weeks, an integer overflow is
triggered, which can cause the neuron to be incorrectly recognized as
inactive. This happened during testing of `claim_neuron`.

This should not affect prod because in prod, `now` should always be much
larger than 2 weeks.

This was discovered using TLA codelink.