feat(ckbtc): rotate KYT RPC providers in case of transient error (#1864)
XC-159: Support multiple RPC providers

Support multiple bitcoin mainnet RPC providers in the KYT canister. Each
RPC call will go to one provider, and the next call will go to the next
provider and so on, in a round robin manner.

When an RPC call returns http error code, it is treated as a transient
error so that the same call can be tried again with the next provider.

Because RPC errors can occur for all providers indefinitely, it is up to
the caller to decide whether or not to call `check_transaction` again.

---------

Co-authored-by: gregorydemay <112856886+gregorydemay@users.noreply.github.com>