feat(PocketIC): add function get_default_effective_canister_id (#2010)
This PR adds the function `get_default_effective_canister_id` to
retrieve a default effective canister id for canister creation on a
PocketIC instance. It takes a URL of the PocketIC instance since the
function is supposed to be used in contexts where a PocketIC instance
(with a potentially unknown instance id) already exists and only a URL
(e.g., of the PocketIC HTTP gateway) is known.

Moreover:
- this PR adds a `default_effective_canister_id` field (as returned by
the function `get_default_effective_canister_id`) into the struct
`Topology`;
- calls to `provisional_create_canister_with_cycles` with the management
canister id as the effective canister id are rejected since this PR
enables to determine a proper effective canister id.