perf: NODE-1499 Don't recursively fix permissions of the `backup` directory after reboots (#1906)
The recent increase in finalization rate has lead to a larger amount of
backup artifacts being retained by the replicas of all mainnet subnets.

The `setup-permissions.sh` script recursively iterates over all files in
the `ic` directory and applies the correct permissions. Due to the
larger amount of backup artifacts, the runtime of this script increased,
leading to a longer downtime during subnet upgrades.

This PR changes the script such that it only sets the permissions of the
top level backup directory. This should be safe because files inside the
backup directory are only written and removed by the replica itself.