fix: return BAD_REQUEST for all validation errors  (#1846)
The PR also logs all validation errors.

[403
(forbidden](https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.3))
really means that the request was good and the server refuses to
process. However, in our case validation error really means that the
request is malformed.

For example, in the RFC it is written 

> If authentication credentials were provided in the request, the server
considers them insufficient to grant access.

In our case it is not about insufficient credential but, malformed
credentials.

---------

Co-authored-by: Martin Raszyk <martin.raszyk@dfinity.org>