feat(IDX): use env vars for embedders test deps (#2047)
This changes the embedders spec tests to use environment variables (set
in `BUILD.bazel`) to specify the path to the wasm spec files.

This ensures that the tests do not depend on Bazel paths like
`external/` and (more theoretically than practically) will make the
tests run only if a file specifically depended on changes.