feat(ICRC-ledger): FI-1532: Check for incompatible downgrade in ICRC ledger (#2019)
If upgrade from V2 to V3 does not complete successfully (the ledger does
not become ready), the ledger can be downgraded to V2 without losing any
data. However, if the upgrade from V2 to V3 is successful, downgrading
back to V2 would result in data loss. All approvals are already in the
stable structures and the serialized/deserialized ledger state would not
contain any approvals.

To avoid the data loss, we introduce ledger versions. V2 is
ledger_version=0, V3 is ledger_version=1. An increase in ledger_version
means an incompatible change that does not allow for downgrade.

---------

Co-authored-by: Mathias Björkqvist <mathias.bjorkqvist@dfinity.org>