chore(crypto): CRP-2574 Use k256 instead of schnorr_fun
In a few areas of the code relating to BIP340 we were previously
forced to use the crate schnorr_fun instead of our usual k256, because
k256 before 0.13.4 did not support BIP340 signatures over messages
that were not exactly 32 bytes.

Now that k256 0.13.4 is available we can use k256 universally and
remove schnorr_fun as a dependency within the monorepo.