refactor(nns): Improve range neurons performance for stable store (#2033)
When scanning neurons in the main StableBTreeMap, instead of doing a
separate traversal of the auxiliary StableBTreeMaps for EACH neuron in
the result set, ranges of each auxiliary StableBTreeMap are used (and
scanned in parallel).

This gets around a 40% performance improvement.

This method, while not currently extensively used, will be used more
after the migration of neurons to stable storage.

Benchmark results before the change:

```
// Output from benchmark test
Benchmark: range_neurons_performance
  total:
    instructions: 47.07 M (improved by 41.47%)
    heap_increase: 0 pages (no change)
    stable_memory_increase: 0 pages (no change)

```

```
range_neurons_performance:
  total:
    instructions: 80415437
    heap_increase: 0
    stable_memory_increase: 0
  scopes: {}
```

After the change (what's in PR):

```
  range_neurons_performance:
    total:
      instructions: 47070657
      heap_increase: 0
      stable_memory_increase: 0
    scopes: {}
```