feat(IDX): Use MODULE.bazel for http archive deps
This ports various http archives to the `MODULE.bazel`, from the
`WORKSPACE`. This covers most `http_archive`s, though focuses on the
ones that pull in external _assets_, and not e.g. rules (will be done
separately).

This builds on previous work to make our repo agnostic to Bazel download
paths:

* https://github.com/dfinity/ic/pull/2047
* https://github.com/dfinity/ic/pull/2062
* https://github.com/dfinity/ic/pull/2096
* https://github.com/dfinity/ic/pull/2097
* https://github.com/dfinity/ic/pull/2102
* https://github.com/dfinity/ic/pull/2126

In the future, and in general, should we avoid introducing new `external/...`
references to make migrations like this much simpler.