ci: Enable interactive testing and inspection of the SetupOS installer.
This new testing feature permits the execution of a SetupOS installer in two modes:

1. Automatic install then reboot to HostOS.
2. Interactive debug without install.

This only works in the dev environment established by Bazel.

This feat is accomplished by the addition of the following variants:

* //ic-os/setupos/envs/dev:interactive-install for installation (with -no-kvm for scenarios where no acceleration exists)
* //ic-os/setupos/envs/dev:interactive-debug for no-install inspection of SetupOS (same -no-kvm availability)

To enable this to work with higher performance, we create a new disk image variant that does not tarball the SetupOS image.
Copying the image to the temporary directory is near instantaneous since various kernel optimizations at play using `cp`
kick in that do not kick in when using `tar x`.

Total time in my ThinkPad (without KVM acceleration) for the whole install + reboot + boot to HostOS + login + poweroff is:

```
real    7m31.450s
user    6m16.889s
sys     0m17.826s
```